Rename existing components to use names without "component" in the filename or component name
